move execution log and execution log entry into contexts
