+++ council/contexts/__init__.py
@@ -2,6 +2,9 @@
 contains the core classes of the engine
 """
 
+from ._execution_log_entry import ExecutionLogEntry
+from ._execution_log import ExecutionLog
+
 from .budget import Consumption, ConsumptionEvent, Budget, InfiniteBudget
 from .cancellation_token import CancellationToken
 from .execution_context import (
