+++ council/controllers/basic_controller.py
@@ -1,7 +1,6 @@
 from typing import List
 
 from council.contexts import AgentContext, Budget
-
 from .controller_base import ControllerBase
 from .execution_unit import ExecutionUnit
